requirements.txt
altair==4.2.0
elasticsearch==7.11.0
gensim==4.3.0
geopandas==0.10.2
ipython==8.8.0
lightgbm==3.3.3
lxml==4.9.2
nltk==3.8.1
numpy==1.23.5
opencv_python==4.7.0.68
pandas==1.5.2
Pillow==9.4.0
py4j==0.10.9.5
pyclustering==0.10.1.2
pydotplus==2.0.2
pyfim==6.28
pyspark==3.3.1
requests==2.28.1
Rtree==1.0.1
scikit_learn==1.2.0
scipy==1.10.0
Scrapy==2.7.1
seaborn==0.12.2
selenium==4.7.2
simplejson==3.18.1
simpletransformers==0.63.9
tensorflow==2.10.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
tweepy==4.12.1
Unidecode==1.3.6
wordcloud==1.8.2.2
XAI_Library==0.0.4
